Introduced variable to specifiy filter setting
Now an integer in the config file indicates if real or reconstructed
events should be used for reconstruction or if cuts should be applied.
This is passed by CCMLEM to the InputReaderGeant. Implementation in
InputReaderGeant is still missing.

Bugfix: Commented out some calculation to alter positions which made
the reconstruction fail and produced empty images